Uploading chapter 1
This commit includes the chapter 1 of the Arduino Twitter Bot project.
In This chapter an Arduino is programmed to connect to a network and
send demo data to a little server which then prints the data in the
console. This builds the foundation to send data collected with the
Arduino to our database. Further details will be added any time soon.