Finalizing Chapter 1
With this commit Chapter 1 now includes the code to read data from a RTC
and a temperature and humidity sensor and sends them to the server.
There are two sub-chapters that deal with how the data is collected and
how it is sent.
These two results are combiend in the file chapter1.ino which collects
the data and sends them to the server.