Uploading Chapter 3
In this chapter the chapters 1 and two are merged together to upload the
data collected by the Arduino to Google Sheets. Mostly this is just a
copy and paste of the previous chapters with some minor tweaks to get it
run. In the wiki soon you can read what changes have been made and why.